refactor(loader): 데이터 로더를 책임지는 클래스들 역할 재분배
- 사용하지 않는 클래스 삭제
