+++ src/main/java/attendance/Application.java
@@ -3,10 +3,9 @@ package attendance;
 import attendance.controller.Controller;
 import attendance.model.Crews;
 import attendance.model.FixedCustomClock;
-import attendance.model.loader.CrewDataLoader;
+import attendance.model.loader.AttendanceLoader;
+import attendance.model.loader.AttendanceProcessor;
 import attendance.model.loader.CrewRegistry;
-import attendance.model.loader.RawAttendanceStore;
-import attendance.service.AttendanceProcessor;
 import attendance.view.InputView;
 import attendance.view.OutputView;
 import java.time.LocalDate;
@@ -18,13 +17,9 @@ public class Application {
     public static void main(String[] args) {
         Crews crews = new Crews(new ArrayList<>());
         FixedCustomClock fixedCustomClock = createCustomClock();
-        RawAttendanceStore rawAttendanceStore = new RawAttendanceStore();
-        CrewDataLoader crewDataLoader = new CrewDataLoader(rawAttendanceStore);
+        AttendanceLoader attendanceLoader = new AttendanceLoader();
         CrewRegistry crewRegistry = new CrewRegistry(crews);
-        AttendanceProcessor attendanceProcessor = new AttendanceProcessor(crewDataLoader, crewRegistry,
-                fixedCustomClock, crews, rawAttendanceStore);
-
-        attendanceProcessor.processAttendanceRecords();
+        loadAttendance(attendanceLoader, fixedCustomClock, crewRegistry);
 
         InputView inputView = new InputView(fixedCustomClock);
         OutputView outputView = new OutputView();
@@ -37,5 +32,14 @@ public class Application {
                 LocalDate.of(2024, 12, 25)
         );
         return new FixedCustomClock(holidays);
+    }
+
+    private static void loadAttendance(AttendanceLoader attendanceLoader, FixedCustomClock fixedCustomClock,
+                                       CrewRegistry crewRegistry) {
+        AttendanceProcessor attendanceProcessor = new AttendanceProcessor(
+                attendanceLoader, fixedCustomClock, crewRegistry
+        );
+
+        attendanceProcessor.processAttendanceRecords();
     }
 }
