+++ src/main/java/attendance/model/loader/CrewRegistry.java
@@ -24,4 +24,3 @@ public class CrewRegistry {
     }
 
 }
-
