--- src/test/java/attendance/TestFixedCustomClock.java
@@ -1,8 +1,0 @@
-package attendance;
-
-//public class TestFixedCustomClock extends FixedCustomClock {
-//    public TestFixedCustomClock() {
-//        super(Set.of(LocalDate.of(2024, 12, 25))); // ✅ 테스트용 공휴일 설정
-//    }
-//
-//}
